% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_image_path.R
\name{find_image_path}
\alias{find_image_path}
\title{Get an image path from a clock}
\usage{
find_image_path(time, folder, clock, path = NULL, github = TRUE)
}
\arguments{
\item{time}{character string in hms format. Which time does the image show?}

\item{folder}{character string showing the clock specific folder format (output from \code{list_times})}

\item{clock}{character string. For which clock is additional information required?}

\item{path}{character string. Path to the Github repo \code{heike/ticktock-images} (or a local clone).}

\item{github}{binary value; true if the give path points to a github repository}
}
\value{
path to an image showing the specified time (based on its name, not the clock hand). In case of a warning the image corresponding to the closest time is shown.
}
\description{
Get the path (or a different example path) to an image for a specific clock.
}
\examples{
# example code

find_image_path("18:21:55", "clock2019-10-16_18", "omni", "heike/ticktock-images", TRUE)
find_image_path("18:59:59", "clock2019-10-16_18", "omni", "heike/ticktock-images", TRUE)
}
