% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Get an image path from a clock}
\usage{
get_image(path, repo = NULL, github = TRUE)
}
\arguments{
\item{path}{character string showing the path to the file in the repo/folder (output from \code{find_image_path})}

\item{repo}{character string. Path to the Github repo \code{heike/ticktock-images} (or a local clone).}

\item{github}{binary value; true if the give path points to a github repository}
}
\value{
path to an image showing the specified time (based on its name, not the clock hand). In case of a warning the image corresponding to the closest time is shown.
}
\description{
Get an image path from a clock
}
\examples{
# example code

im <- get_image("omni/clock2019-10-16_18/clock2019-10-16_18.59.59.jpg",
"heike/ticktock-images", TRUE)
plot(im)
im <- get_image("omni/clock2019-10-16_18/clock2019-10-16_18.29.15.jpg",
"heike/ticktock-images", TRUE)
plot(im)
}
