% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\title{Get an image based on time}
\usage{
get_time(time, sample_path = NULL, repo = NULL, github = TRUE)
}
\arguments{
\item{time}{character string in the form hour, minute, second.}

\item{sample_path}{character string showing the path to a file in the repo/folder (output from \code{find_image_path}), marked up with \verb{\%HOUR\%}, \verb{\%MINUTE\%} and \verb{\%SECOND\%} in appropriate places.}

\item{repo}{character string. Path to the Github repo \code{heike/ticktock-images} (or a local clone).}

\item{github}{binary value; true if the give path points to a github repository}
}
\value{
path to an image showing the specified time (based on its name, not the clock hand). In case of a warning the image corresponding to the closest time is shown.
}
\description{
A timestamp will be inserted into the correct format to pull an existing image from a repository.
}
\examples{
# example code

im <- get_time("18:15:39",
"omni/clock2019-10-16_\%HOUR\%/clock2019-10-16_\%HOUR\%.\%MINUTE\%.\%SECOND\%.jpg",
"heike/ticktock-images", TRUE)
plot(im)
im <- get_time("18:15:44",
"omni/clock2019-10-16_\%HOUR\%/clock2019-10-16_\%HOUR\%.\%MINUTE\%.\%SECOND\%.jpg",
"heike/ticktock-images", TRUE)
# careful - the image for the next second does not exist.
plot(im)
}
