% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_clocks.R
\name{list_clocks}
\alias{list_clocks}
\title{Get a list of all available recordings by clock}
\usage{
list_clocks(path, github = TRUE)
}
\arguments{
\item{path}{character string. Path to the Github repo \code{heike/ticktock-images} (or a local clone).}

\item{github}{binary value; true if the give path points to a github repository}
}
\value{
vector of available clock recordings
}
\description{
Get a list of all available recordings by clock
}
\examples{
# example code
list_clocks("heike/ticktock-images")
}
